Removed cast, which is now superfluous.

+++ samples/seissol/proxy_seissol_allocator.hpp
@@ -242,9 +242,9 @@ unsigned int init_data_structures(unsigned int i_cells) {
         m_cellInformation[l_cell].faceNeighborIds[f] =  scenario_neighbor[l_cell][f];
       } else {
         m_cellInformation[l_cell].faceTypes[f] = regular;
-        m_cellInformation[l_cell].faceRelations[f][0] = ((unsigned int)libxsmm_rand_u32(4));
-        m_cellInformation[l_cell].faceRelations[f][1] = ((unsigned int)libxsmm_rand_u32(3));
-        m_cellInformation[l_cell].faceNeighborIds[f] = ((unsigned int)libxsmm_rand_u32(i_cells));
+        m_cellInformation[l_cell].faceRelations[f][0] = (libxsmm_rand_u32(4));
+        m_cellInformation[l_cell].faceRelations[f][1] = (libxsmm_rand_u32(3));
+        m_cellInformation[l_cell].faceNeighborIds[f] = (libxsmm_rand_u32(i_cells));
       }
     }
 #ifdef __USE_DERS
